hw5/HW5-3180101941-Yangrui/cnn/requirements.txt
numpy >= 1.8.1
matplotlib.pyplot >= 3.1.2
tensorflow >= 2.2.0